refactor(agent-bff): extract execute validation tests and shared fixtures
