+++ packages/agent-bff/test/action/action-routes-middleware.test.ts
@@ -1,6 +1,5 @@
-import type { Action, AgentActionClient } from '../../src/action/agent-action-client';
+import type { AgentActionClient } from '../../src/action/agent-action-client';
 import type { Logger } from '../../src/ports/logger-port';
-import type ReadModelStore from '../../src/read-model/read-model-store';
 
 import {
   ActionFormValidationError,
@@ -16,161 +15,16 @@ import request from 'supertest';
 import createActionRoutesMiddleware from '../../src/action/action-routes-middleware';
 import createErrorMiddleware from '../../src/http/error-middleware';
 import SchemaUnavailableError from '../../src/read-model/errors';
-import ReadModel from '../../src/read-model/read-model';
-import { action, collection, column } from '../read-model/fixtures';
-
-const TIMEZONE = 'Europe/Paris';
-
-const noopLogger: Logger = () => {};
-
-function storeOf(readModel: ReadModel | Error): ReadModelStore {
-  return {
-    getReadModel: async () => {
-      if (readModel instanceof Error) throw readModel;
-
-      return readModel;
-    },
-  } as unknown as ReadModelStore;
-}
-
-interface FakeField {
-  name: string;
-  type: string | [string];
-  value: unknown;
-  isRequired: boolean;
-  enumValues?: string[];
-  reference?: string | null;
-}
-
-function makeAction({
-  fields = [],
-  layout = [],
-  skipped = [],
-  postChange,
-  postSet,
-  execute = jest.fn(),
-  setFields,
-}: {
-  fields?: FakeField[];
-  layout?: unknown[];
-  skipped?: string[];
-  postChange?: { fields?: FakeField[]; layout?: unknown[] };
-  postSet?: { fields?: FakeField[]; layout?: unknown[] };
-  execute?: jest.Mock;
-  setFields?: jest.Mock;
-} = {}) {
-  const state = { fields: [...fields], layout: [...layout] };
-
-  // Mirrors the real Action: setFields assigns each submitted value onto the matching field
-  // (unknown fields would throw there; tests that need that pass their own setFields mock), and a
-  // change hook can rebuild the fields in place, which is what a postSet fixture simulates.
-  const applyValues = async (values: Record<string, unknown>) => {
-    for (const [name, value] of Object.entries(values)) {
-      const field = state.fields.find(f => f.name === name);
-      if (field) field.value = value;
-    }
-
-    if (postSet) {
-      state.fields = postSet.fields ?? state.fields;
-      state.layout = postSet.layout ?? state.layout;
-    }
-  };
-
-  const tryToSetFields = jest.fn(async () => {
-    if (postChange) {
-      state.fields = postChange.fields ?? state.fields;
-      state.layout = postChange.layout ?? state.layout;
-    }
-
-    return skipped;
-  });
-
-  const form: Action & { tryToSetFields: jest.Mock; execute: jest.Mock; setFields: jest.Mock } = {
-    execute,
-    setFields: setFields ?? jest.fn(applyValues),
-    tryToSetFields,
-    getFields: () =>
-      state.fields.map(f => ({
-        getName: () => f.name,
-        getType: () => f.type,
-        getReference: () => f.reference ?? null,
-        getValue: () => f.value,
-        isRequired: () => f.isRequired,
-      })),
-    getEnumField: (name: string) => ({
-      getOptions: () => state.fields.find(f => f.name === name)?.enumValues,
-    }),
-    getLayout: () => ({ layout: state.layout }),
-  };
-
-  return form;
-}
-
-function clientOf(
-  form: Action,
-  loadAction: jest.Mock = jest.fn(async () => form),
-): AgentActionClient & { loadAction: jest.Mock } {
-  return { loadAction };
-}
-
-function buildApp(
-  store: ReadModelStore,
-  client: AgentActionClient,
-  {
-    agentToken = 'agent-jwt',
-    logger = noopLogger,
-    timezone = TIMEZONE,
-  }: { agentToken?: string | null; logger?: Logger; timezone?: string } = {},
-) {
-  const app = new Koa();
-  app.silent = true;
-  app.use(createErrorMiddleware({ logger: noopLogger }));
-  app.use(bodyParser());
-  app.use(async (ctx, next) => {
-    ctx.state.timezone = timezone;
-    if (agentToken !== null) ctx.state.agentToken = agentToken;
-    await next();
-  });
-  app.use(
-    createActionRoutesMiddleware({
-      store,
-      agentUrl: 'https://agent.example.com',
-      logger,
-      createClient: () => client,
-    }),
-  );
-
-  return app;
-}
-
-const readModel = new ReadModel([
-  collection('users', [column('id')], [action('approve', '/forest/_actions/users/0/approve')]),
-]);
-
-function buildAppWithTerminal(client: AgentActionClient) {
-  const app = new Koa();
-  app.silent = true;
-  app.use(createErrorMiddleware({ logger: noopLogger }));
-  app.use(bodyParser());
-  app.use(async (ctx, next) => {
-    ctx.state.timezone = TIMEZONE;
-    ctx.state.agentToken = 'agent-jwt';
-    await next();
-  });
-  app.use(
-    createActionRoutesMiddleware({
-      store: storeOf(readModel),
-      agentUrl: 'https://agent.example.com',
-      logger: noopLogger,
-      createClient: () => client,
-    }),
-  );
-  app.use(async ctx => {
-    ctx.status = 204;
-  });
-
-  return app;
-}
+import {
+  TIMEZONE,
+  buildApp,
+  buildAppWithTerminal,
+  clientOf,
+  makeAction,
+  noopLogger,
+  readModel,
+  storeOf,
+} from '../helpers/action-routes';
 
 describe('action routes middleware', () => {
   it('forwards the configured agent timeout to the action client', async () => {
@@ -842,412 +696,6 @@ describe('action execute', () => {
     expect(response.status).toBe(502);
     expect(response.body.error).toMatchObject({ type: 'network_error', status: 502 });
     expect(execute).not.toHaveBeenCalled();
-  });
-
-  describe('execute input validation', () => {
-    it('rejects a missing required field with 400 missing_required_fields before executing', async () => {
-      const execute = jest.fn(async () => ({ success: 'Done' }));
-      const form = makeAction({
-        fields: [
-          { name: 'reason', type: 'String', value: null, isRequired: true },
-          { name: 'tier', type: 'String', value: null, isRequired: true },
-        ],
-        execute,
-      });
-
-      const response = await request(execApp(clientOf(form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .send({ recordIds: ['42'], values: {} });
-
-      expect(response.status).toBe(400);
-      expect(response.body.error).toEqual({
-        type: 'missing_required_fields',
-        status: 400,
-        message: 'Required action fields are missing: reason, tier',
-        details: { fields: ['reason', 'tier'] },
-      });
-      expect(execute).not.toHaveBeenCalled();
-    });
-
-    it('executes when the submitted value satisfies the required field', async () => {
-      const execute = jest.fn(async () => ({ success: 'Done' }));
-      const form = makeAction({
-        fields: [
-          { name: 'reason', type: 'String', value: null, isRequired: true },
-          // 0 and false count as present values, exactly like the form endpoint's requiredFields.
-          { name: 'amount', type: 'Number', value: null, isRequired: true },
-          { name: 'confirm', type: 'Boolean', value: null, isRequired: true },
-        ],
-        execute,
-      });
-
-      const response = await request(execApp(clientOf(form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .send({ recordIds: ['42'], values: { reason: 'because', amount: 0, confirm: false } });
-
-      expect(response.status).toBe(200);
-      expect(execute).toHaveBeenCalledTimes(1);
-    });
-
-    it('executes when a required field is satisfied by a loaded default, not a submission', async () => {
-      const execute = jest.fn(async () => ({ success: 'Done' }));
-      const form = makeAction({
-        fields: [{ name: 'reason', type: 'String', value: 'preset', isRequired: true }],
-        execute,
-      });
-
-      const response = await request(execApp(clientOf(form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .send({ recordIds: ['42'], values: {} });
-
-      expect(response.status).toBe(200);
-      expect(execute).toHaveBeenCalledTimes(1);
-    });
-
-    it('validates the form as rebuilt by a change hook, not the loaded one', async () => {
-      const execute = jest.fn(async () => ({ success: 'Done' }));
-      // The loaded form requires nothing; the hook that runs on setFields makes "late" required.
-      const form = makeAction({
-        fields: [{ name: 'reason', type: 'String', value: null, isRequired: false }],
-        postSet: { fields: [{ name: 'late', type: 'String', value: null, isRequired: true }] },
-        execute,
-      });
-
-      const response = await request(execApp(clientOf(form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .send({ recordIds: ['42'], values: { reason: 'x' } });
-
-      expect(response.status).toBe(400);
-      expect(response.body.error).toMatchObject({
-        type: 'missing_required_fields',
-        details: { fields: ['late'] },
-      });
-      expect(execute).not.toHaveBeenCalled();
-    });
-
-    it('rejects an out-of-enum value with 422 invalid_action_value before executing', async () => {
-      const execute = jest.fn(async () => ({ success: 'Done' }));
-      const form = makeAction({
-        fields: [
-          {
-            name: 'tier',
-            type: 'Enum',
-            value: null,
-            isRequired: false,
-            enumValues: ['gold', 'silver'],
-          },
-        ],
-        execute,
-      });
-
-      const response = await request(execApp(clientOf(form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .send({ recordIds: ['42'], values: { tier: 'NOPE' } });
-
-      expect(response.status).toBe(422);
-      expect(response.body.error).toEqual({
-        type: 'invalid_action_value',
-        status: 422,
-        message: 'Invalid action field values: tier (expected one of: gold, silver)',
-        details: { fields: [{ field: 'tier', expected: 'one of: gold, silver' }] },
-      });
-      expect(execute).not.toHaveBeenCalled();
-    });
-
-    it('executes an in-options enum value, a number on a Number field, and an enum without options', async () => {
-      const execute = jest.fn(async () => ({ success: 'Done' }));
-      const form = makeAction({
-        fields: [
-          { name: 'tier', type: 'Enum', value: null, isRequired: false, enumValues: ['gold'] },
-          // An agent can declare an Enum with no options: the field then accepts any string.
-          { name: 'free', type: 'Enum', value: null, isRequired: false, enumValues: [] },
-          { name: 'amount', type: 'Number', value: null, isRequired: false },
-        ],
-        execute,
-      });
-
-      const response = await request(execApp(clientOf(form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .send({ recordIds: ['42'], values: { tier: 'gold', free: 'anything', amount: 5 } });
-
-      expect(response.status).toBe(200);
-      expect(execute).toHaveBeenCalledTimes(1);
-    });
-
-    it('rejects a null item inside a constrained list before executing', async () => {
-      const makeListForm = (values: Record<string, unknown>) => {
-        const execute = jest.fn(async () => ({ success: 'Done' }));
-
-        return {
-          execute,
-          response: request(
-            execApp(
-              clientOf(
-                makeAction({
-                  fields: [
-                    {
-                      name: 'tags',
-                      type: ['Enum'],
-                      value: null,
-                      isRequired: false,
-                      enumValues: ['a'],
-                    },
-                    { name: 'ids', type: ['Number'], value: null, isRequired: false },
-                  ],
-                  execute,
-                }),
-              ),
-            ).callback(),
-          )
-            .post('/agent/v1/users/actions/approve/execute')
-            .send({ recordIds: ['42'], values }),
-        };
-      };
-
-      const enumResult = await makeListForm({ tags: ['a', null] }).response;
-      expect(enumResult.status).toBe(422);
-      expect(enumResult.body.error).toMatchObject({
-        type: 'invalid_action_value',
-        details: { fields: [{ field: 'tags', expected: 'one of: a' }] },
-      });
-
-      const numberResult = await makeListForm({ ids: [1, null] }).response;
-      expect(numberResult.status).toBe(422);
-      expect(numberResult.body.error).toMatchObject({
-        type: 'invalid_action_value',
-        details: { fields: [{ field: 'ids', expected: 'a number' }] },
-      });
-    });
-
-    it('rejects a form-urlencoded execute body: it cannot carry the nested recordIds and values', async () => {
-      // A urlencoded form flattens bracket notation instead of nesting it, so recordIds never
-      // arrives as an array and the request fails at the shape check, before value validation.
-      const execute = jest.fn(async () => ({ success: 'Done' }));
-      const form = makeAction({
-        fields: [{ name: 'amount', type: 'Number', value: null, isRequired: false }],
-        execute,
-      });
-
-      const response = await request(execApp(clientOf(form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .type('form')
-        .send({ recordIds: ['42'], values: { amount: '5' } });
-
-      expect(response.status).toBe(400);
-      expect(response.body.error).toMatchObject({ type: 'invalid_request', status: 400 });
-      expect(execute).not.toHaveBeenCalled();
-    });
-
-    it.each([
-      ['a string sent to a Number field', 'amount', 'Number', 'nan', 'a number'],
-      ['a string sent to a Boolean field', 'flag', 'Boolean', 'true', 'a boolean'],
-      ['a number sent to a String field', 'reason', 'String', 42, 'a string'],
-    ])(
-      'rejects %s with 422 naming the expectation',
-      async (_label, name, type, value, expected) => {
-        const execute = jest.fn(async () => ({ success: 'Done' }));
-        const form = makeAction({
-          fields: [{ name, type, value: null, isRequired: false }],
-          execute,
-        });
-
-        const response = await request(execApp(clientOf(form)).callback())
-          .post('/agent/v1/users/actions/approve/execute')
-          .send({ recordIds: ['42'], values: { [name]: value } });
-
-        expect(response.status).toBe(422);
-        expect(response.body.error).toEqual({
-          type: 'invalid_action_value',
-          status: 422,
-          message: `Invalid action field values: ${name} (expected ${expected})`,
-          details: { fields: [{ field: name, expected }] },
-        });
-        expect(execute).not.toHaveBeenCalled();
-      },
-    );
-
-    it('rejects a non-array sent to a list field and an out-of-options list item', async () => {
-      const makeListForm = () => {
-        const execute = jest.fn(async () => ({ success: 'Done' }));
-
-        return {
-          execute,
-          form: makeAction({
-            fields: [
-              { name: 'ids', type: ['Number'], value: null, isRequired: false },
-              {
-                name: 'tags',
-                type: ['Enum'],
-                value: null,
-                isRequired: false,
-                enumValues: ['a', 'b'],
-              },
-            ],
-            execute,
-          }),
-        };
-      };
-
-      // Each request loads a fresh Action from the agent, so the two cases get separate forms.
-      const firstCase = makeListForm();
-      const first = await request(execApp(clientOf(firstCase.form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .send({ recordIds: ['42'], values: { ids: 'nope' } });
-
-      expect(first.status).toBe(422);
-      expect(first.body.error).toMatchObject({
-        type: 'invalid_action_value',
-        details: { fields: [{ field: 'ids', expected: 'an array of a number' }] },
-      });
-
-      const secondCase = makeListForm();
-      const second = await request(execApp(clientOf(secondCase.form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .send({ recordIds: ['42'], values: { tags: ['a', 'zzz'] } });
-
-      expect(second.status).toBe(422);
-      expect(second.body.error).toMatchObject({
-        type: 'invalid_action_value',
-        details: { fields: [{ field: 'tags', expected: 'one of: a, b' }] },
-      });
-      expect(firstCase.execute).not.toHaveBeenCalled();
-      expect(secondCase.execute).not.toHaveBeenCalled();
-    });
-
-    it('executes a fully valid list submission', async () => {
-      const execute = jest.fn(async () => ({ success: 'Done' }));
-      const form = makeAction({
-        fields: [
-          { name: 'ids', type: ['Number'], value: null, isRequired: false },
-          {
-            name: 'tags',
-            type: ['Enum'],
-            value: null,
-            isRequired: false,
-            enumValues: ['a', 'b'],
-          },
-        ],
-        execute,
-      });
-
-      const response = await request(execApp(clientOf(form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .send({ recordIds: ['42'], values: { ids: [1, 2], tags: ['a', 'b'] } });
-
-      expect(response.status).toBe(200);
-      expect(execute).toHaveBeenCalledTimes(1);
-    });
-
-    it('does not type-check Json or File fields at the BFF', async () => {
-      const execute = jest.fn(async () => ({ success: 'Done' }));
-      const form = makeAction({
-        fields: [
-          { name: 'meta', type: 'Json', value: null, isRequired: false },
-          { name: 'doc', type: 'File', value: null, isRequired: false },
-        ],
-        execute,
-      });
-
-      const response = await request(execApp(clientOf(form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .send({
-          recordIds: ['42'],
-          values: { meta: { any: [1, null] }, doc: 'data:text/plain;aGk=' },
-        });
-
-      expect(response.status).toBe(200);
-      expect(execute).toHaveBeenCalledTimes(1);
-    });
-
-    it('accepts a packed id string on a record-picker field whose PK type is Number', async () => {
-      // The agent rewrites a Collection field to its target PK column type, but execute unpacks
-      // packed id STRINGS only (IdUtils.unpackId throws on a number): the published type is not
-      // the runtime contract, and the string is the only shape that may execute.
-      const execute = jest.fn(async () => ({ success: 'Done' }));
-      const form = makeAction({
-        fields: [
-          {
-            name: 'assignee',
-            type: 'Number',
-            value: null,
-            isRequired: false,
-            reference: 'users.id',
-          },
-        ],
-        execute,
-      });
-
-      const ok = await request(execApp(clientOf(form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .send({ recordIds: ['42'], values: { assignee: '42' } });
-
-      expect(ok.status).toBe(200);
-      expect(execute).toHaveBeenCalledTimes(1);
-    });
-
-    it('rejects a raw number on a record-picker field with the packed id expectation', async () => {
-      const execute = jest.fn(async () => ({ success: 'Done' }));
-      const form = makeAction({
-        fields: [
-          {
-            name: 'assignee',
-            type: 'Number',
-            value: null,
-            isRequired: false,
-            reference: 'users.id',
-          },
-        ],
-        execute,
-      });
-
-      const response = await request(execApp(clientOf(form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .send({ recordIds: ['42'], values: { assignee: 42 } });
-
-      expect(response.status).toBe(422);
-      expect(response.body.error).toEqual({
-        type: 'invalid_action_value',
-        status: 422,
-        message:
-          'Invalid action field values: assignee (expected a string holding a packed record id)',
-        details: {
-          fields: [{ field: 'assignee', expected: 'a string holding a packed record id' }],
-        },
-      });
-      expect(execute).not.toHaveBeenCalled();
-    });
-
-    it('ignores a leftover enums array on a typed field and a malformed options array', async () => {
-      const execute = jest.fn(async () => ({ success: 'Done' }));
-      const form = makeAction({
-        fields: [
-          // A hook can rebuild a form leaving enums on a field whose type is not Enum: membership
-          // applies to Enum fields only. `enums` arrives malformed (a bare string) on the other.
-          {
-            name: 'amount',
-            type: 'Number',
-            value: null,
-            isRequired: false,
-            enumValues: ['gold', 'silver'],
-          },
-          {
-            name: 'tier',
-            type: 'Enum',
-            value: null,
-            isRequired: false,
-            enumValues: 'gold' as unknown as string[],
-          },
-        ],
-        execute,
-      });
-
-      const response = await request(execApp(clientOf(form)).callback())
-        .post('/agent/v1/users/actions/approve/execute')
-        .send({ recordIds: ['42'], values: { amount: 5, tier: 'whatever' } });
-
-      expect(response.status).toBe(200);
-      expect(execute).toHaveBeenCalledTimes(1);
-    });
   });
 
   it('maps an approval-gated action to 403 action_requires_approval with the allowed roles', async () => {
